# WSGmed/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation",
        "description": "Centralized API documentation",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/api/patients": {
            "get": {
                "tags": [
                    "Patients"
                ],
                "summary": "Get all patients",
                "description": "Retrieves a list of all patients in the system",
                "operationId": "getPatientsList",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Jan"
                                            },
                                            "s_name": {
                                                "type": "string",
                                                "example": "Kowalski"
                                            },
                                            "email": {
                                                "type": "string",
                                                "format": "email",
                                                "example": "jan.kowalski@example.com"
                                            },
                                            "date_of_birth": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "1990-01-15"
                                            },
                                            "location_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-05T12:00:00Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "date-time",
                                                "example": "2025-05-05T12:00:00Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Patients"
                ],
                "summary": "Create a new patient",
                "description": "Creates a new patient with the provided information",
                "operationId": "storePatient",
                "requestBody": {
                    "description": "Patient creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "s_name",
                                    "email",
                                    "date_of_birth",
                                    "location_id",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "Patient's first name",
                                        "type": "string",
                                        "example": "Anna"
                                    },
                                    "s_name": {
                                        "description": "Patient's surname",
                                        "type": "string",
                                        "example": "Nowak"
                                    },
                                    "email": {
                                        "description": "Patient's email address",
                                        "type": "string",
                                        "format": "email",
                                        "example": "anna.nowak@example.com"
                                    },
                                    "date_of_birth": {
                                        "description": "Patient's date of birth",
                                        "type": "string",
                                        "format": "date",
                                        "example": "1995-05-20"
                                    },
                                    "location_id": {
                                        "description": "Location ID where the patient is registered",
                                        "type": "integer",
                                        "example": 2
                                    },
                                    "password": {
                                        "description": "Patient's account password",
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Patient created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "name": {
                                            "type": "string",
                                            "example": "Anna"
                                        },
                                        "s_name": {
                                            "type": "string",
                                            "example": "Nowak"
                                        },
                                        "email": {
                                            "type": "string",
                                            "format": "email",
                                            "example": "anna.nowak@example.com"
                                        },
                                        "date_of_birth": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "1995-05-20"
                                        },
                                        "location_id": {
                                            "type": "integer",
                                            "example": 2
                                        },
                                        "created_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-05T12:00:00Z"
                                        },
                                        "updated_at": {
                                            "type": "string",
                                            "format": "date-time",
                                            "example": "2025-05-05T12:00:00Z"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The given data was invalid"
                                        },
                                        "errors": {
                                            "properties": {
                                                "email": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string",
                                                        "example": "The email has already been taken"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter your JWT token in the format: Bearer {token}"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}